ext: ecsobserver Move ecsmock to internal
It should get moved into internal/aws in the future. Keep it under
extension package for now
